rename formattedDate util, update format userManagement
