+++ src/components/UserManagement/UserTableData.jsx
@@ -10,6 +10,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCopy } from '@fortawesome/free-solid-svg-icons';
 import { toast } from 'react-toastify';
 import { boxStyle } from 'styles';
+import { formattedDate } from 'utils/formattedDate';
 
 /**
  * The body row of the user table
@@ -95,10 +96,10 @@ const UserTableData = React.memo(props => {
       </td>
       <td>
         {props.user.isActive === false && props.user.reactivationDate
-          ? props.user.reactivationDate.toLocaleString().split('T')[0]
+          ? formattedDate(props.user.reactivationDate)
           : ''}
       </td>
-      <td>{props.user.endDate ? props.user.endDate.toLocaleString().split('T')[0] : 'N/A'}</td>
+      <td>{props.user.endDate ? formattedDate(props.user.endDate) : 'N/A'}</td>
       {checkPermissionsOnOwner() ? null : (
         <td>
           <span className="usermanagement-actions-cell">
